+++ tanjun/commands/slash.py
@@ -219,6 +219,7 @@ def as_slash_command(
     dm_enabled: typing.Optional[bool] = None,
     is_global: bool = True,
     sort_options: bool = True,
+    validate_arg_keys: bool = True,
 ) -> _ResultProto:
     r"""Build a [tanjun.SlashCommand][] by decorating a function.
 
@@ -282,6 +283,8 @@ def as_slash_command(
         If this is [True][] then the options are re-sorted to meet the requirement
         from Discord that required command options be listed before optional
         ones.
+    validate_arg_keys
+        Whether to validate that option keys match the command callback's signature.
 
     Returns
     -------
@@ -305,18 +308,11 @@ def as_slash_command(
 
     def decorator(callback: _CallbackishT[_CommandCallbackSigT], /) -> SlashCommand[_CommandCallbackSigT]:
         if isinstance(callback, (tanjun.MenuCommand, tanjun.MessageCommand, tanjun.SlashCommand)):
-            return SlashCommand(
-                callback.callback,
-                name,
-                description,
-                always_defer=always_defer,
-                default_member_permissions=default_member_permissions,
-                default_to_ephemeral=default_to_ephemeral,
-                dm_enabled=dm_enabled,
-                is_global=is_global,
-                sort_options=sort_options,
-                _wrapped_command=callback,
-            )
+            wrapped_command = callback
+            callback = callback.callback
+
+        else:
+            wrapped_command = None
 
         return SlashCommand(
             callback,
@@ -328,6 +324,8 @@ def as_slash_command(
             dm_enabled=dm_enabled,
             is_global=is_global,
             sort_options=sort_options,
+            validate_arg_keys=validate_arg_keys,
+            _wrapped_command=wrapped_command,
         )
 
     return decorator
@@ -1180,6 +1178,7 @@ class SlashCommand(BaseSlashCommand, tanjun.SlashCommand[_CommandCallbackSigT]):
 
     __slots__ = (
         "_always_defer",
+        "_arg_names",
         "_builder",
         "_callback",
         "_client",
@@ -1204,6 +1203,7 @@ class SlashCommand(BaseSlashCommand, tanjun.SlashCommand[_CommandCallbackSigT]):
         dm_enabled: typing.Optional[bool] = None,
         is_global: bool = True,
         sort_options: bool = True,
+        validate_arg_keys: bool = True,
         _wrapped_command: typing.Optional[tanjun.ExecutableCommand[typing.Any]] = None,
     ) -> None:
         ...
@@ -1222,6 +1222,7 @@ class SlashCommand(BaseSlashCommand, tanjun.SlashCommand[_CommandCallbackSigT]):
         dm_enabled: typing.Optional[bool] = None,
         is_global: bool = True,
         sort_options: bool = True,
+        validate_arg_keys: bool = True,
         _wrapped_command: typing.Optional[tanjun.ExecutableCommand[typing.Any]] = None,
     ) -> None:
         ...
@@ -1239,6 +1240,7 @@ class SlashCommand(BaseSlashCommand, tanjun.SlashCommand[_CommandCallbackSigT]):
         dm_enabled: typing.Optional[bool] = None,
         is_global: bool = True,
         sort_options: bool = True,
+        validate_arg_keys: bool = True,
         _wrapped_command: typing.Optional[tanjun.ExecutableCommand[typing.Any]] = None,
     ) -> None:
         r"""Initialise a slash command.
@@ -1298,6 +1300,8 @@ class SlashCommand(BaseSlashCommand, tanjun.SlashCommand[_CommandCallbackSigT]):
             If this is [True][] then the options are re-sorted to meet the requirement
             from Discord that required command options be listed before optional
             ones.
+        validate_arg_keys
+            Whether to validate that option keys match the command callback's signature.
 
         Raises
         ------
@@ -1320,6 +1324,7 @@ class SlashCommand(BaseSlashCommand, tanjun.SlashCommand[_CommandCallbackSigT]):
             callback = callback.callback
 
         self._always_defer = always_defer
+        self._arg_names = utilities.get_kwargs(callback) if validate_arg_keys else None
         self._builder = _SlashCommandBuilder(name, description, sort_options)
         self._callback: _CommandCallbackSigT = callback
         self._client: typing.Optional[tanjun.Client] = None
@@ -1410,7 +1415,6 @@ class SlashCommand(BaseSlashCommand, tanjun.SlashCommand[_CommandCallbackSigT]):
         pass_as_kwarg: bool = True,
         _stack_level: int = 0,
     ) -> _SlashCommandT:
-        _validate_name(name)
         if len(description) > 100:
             raise ValueError("The option description cannot be over 100 characters in length")
 
@@ -1419,6 +1423,10 @@ class SlashCommand(BaseSlashCommand, tanjun.SlashCommand[_CommandCallbackSigT]):
 
         if min_value and max_value and min_value > max_value:
             raise ValueError("The min value cannot be greater than the max value")
+
+        _validate_name(name)
+        if self._arg_names is not None and name not in self._arg_names:
+            raise ValueError(f"{name!r} is not a valid keyword argument for {self._callback}")
 
         type_ = hikari.OptionType(type_)
         if isinstance(converters, collections.Iterable):
@@ -1528,6 +1536,8 @@ class SlashCommand(BaseSlashCommand, tanjun.SlashCommand[_CommandCallbackSigT]):
             * If the option name has uppercase characters.
             * If the option description is over 100 characters in length.
             * If the command already has 25 options.
+            * If `name` isn't valid for this command's callback when
+              `validate_arg_keys` is [True][].
         """
         return self._add_option(
             name,
@@ -1615,6 +1625,8 @@ class SlashCommand(BaseSlashCommand, tanjun.SlashCommand[_CommandCallbackSigT]):
             * If the option description is over 100 characters in length.
             * If the option has more than 25 choices.
             * If the command already has 25 options.
+            * If `name` isn't valid for this command's callback when
+              `validate_arg_keys` is [True][].
         """
         if choices is None:
             actual_choices = None
@@ -1736,6 +1748,8 @@ class SlashCommand(BaseSlashCommand, tanjun.SlashCommand[_CommandCallbackSigT]):
             * If the option has more than 25 choices.
             * If the command already has 25 options.
             * If `min_value` is greater than `max_value`.
+            * If `name` isn't valid for this command's callback when
+              `validate_arg_keys` is [True][].
         """
         self._add_option(
             name,
@@ -1842,6 +1856,8 @@ class SlashCommand(BaseSlashCommand, tanjun.SlashCommand[_CommandCallbackSigT]):
             * If the option has more than 25 choices.
             * If the command already has 25 options.
             * If `min_value` is greater than `max_value`.
+            * If `name` isn't valid for this command's callback when
+              `validate_arg_keys` is [True][].
         """
         self._add_option(
             name,
@@ -1909,6 +1925,8 @@ class SlashCommand(BaseSlashCommand, tanjun.SlashCommand[_CommandCallbackSigT]):
             * If the option name has uppercase characters.
             * If the option description is over 100 characters in length.
             * If the command already has 25 options.
+            * If `name` isn't valid for this command's callback when
+              `validate_arg_keys` is [True][].
         """
         return self._add_option(
             name, description, hikari.OptionType.BOOLEAN, default=default, pass_as_kwarg=pass_as_kwarg
@@ -1966,6 +1984,8 @@ class SlashCommand(BaseSlashCommand, tanjun.SlashCommand[_CommandCallbackSigT]):
             * If the option description is over 100 characters in length.
             * If the option has more than 25 choices.
             * If the command already has 25 options.
+            * If `name` isn't valid for this command's callback when
+              `validate_arg_keys` is [True][].
         """
         return self._add_option(name, description, hikari.OptionType.USER, default=default, pass_as_kwarg=pass_as_kwarg)
 
@@ -2018,6 +2038,8 @@ class SlashCommand(BaseSlashCommand, tanjun.SlashCommand[_CommandCallbackSigT]):
             * If the option name has uppercase characters.
             * If the option description is over 100 characters in length.
             * If the command already has 25 options.
+            * If `name` isn't valid for this command's callback when
+              `validate_arg_keys` is [True][].
         """
         return self._add_option(name, description, hikari.OptionType.USER, default=default, only_member=True)
 
@@ -2077,6 +2099,8 @@ class SlashCommand(BaseSlashCommand, tanjun.SlashCommand[_CommandCallbackSigT]):
             * If the option description is over 100 characters in length.
             * If the command already has 25 options.
             * If an invalid type is passed in `types`.
+            * If `name` isn't valid for this command's callback when
+              `validate_arg_keys` is [True][].
         """
         import itertools
 
@@ -2145,6 +2169,8 @@ class SlashCommand(BaseSlashCommand, tanjun.SlashCommand[_CommandCallbackSigT]):
             * If the option name has uppercase characters.
             * If the option description is over 100 characters in length.
             * If the command already has 25 options.
+            * If `name` isn't valid for this command's callback when
+              `validate_arg_keys` is [True][].
         """
         return self._add_option(name, description, hikari.OptionType.ROLE, default=default, pass_as_kwarg=pass_as_kwarg)
 
@@ -2198,6 +2224,8 @@ class SlashCommand(BaseSlashCommand, tanjun.SlashCommand[_CommandCallbackSigT]):
             * If the option name has uppercase characters.
             * If the option description is over 100 characters in length.
             * If the command already has 25 options.
+            * If `name` isn't valid for this command's callback when
+              `validate_arg_keys` is [True][].
         """
         return self._add_option(
             name, description, hikari.OptionType.MENTIONABLE, default=default, pass_as_kwarg=pass_as_kwarg