+++ tanjun/utilities.py
@@ -45,6 +45,7 @@ __all__: list[str] = [
 ]
 
 import asyncio
+import inspect
 import typing
 from collections import abc as collections
 
@@ -456,3 +457,27 @@ class CastedView(collections.Mapping[_KeyT, _OtherValueT], typing.Generic[_KeyT,
 
     def __len__(self) -> int:
         return len(self._raw_data)
+
+
+_KEYWORD_TYPES = {inspect.Parameter.KEYWORD_ONLY, inspect.Parameter.POSITIONAL_OR_KEYWORD}
+
+
+def get_kwargs(callback: collections.Callable[..., typing.Any]) -> list[str] | None:
+    names: list[str] = []
+
+    try:
+        signature = inspect.Signature.from_callable(callback)
+
+    except ValueError:
+        # When "no signature [is] found" for a callback/type, we just don't
+        # know what parameters it has so we have to assume var keyword.
+        return None
+
+    for parameter in signature.parameters.values():
+        if parameter.kind is parameter.VAR_KEYWORD:
+            return None
+
+        if parameter.kind in _KEYWORD_TYPES:
+            names.append(parameter.name)
+
+    return names